requirements.txt
streamlit==1.37.1
python-dotenv==1.0.1
agno==1.7.11
youtube-transcript-api==0.6.2